Unified error handling focusing on REST Api
