--- ecm-client-backend/src/main/java/com/avispa/ecm/util/api/exception/ApiException.java
@@ -1,40 +1,0 @@
-/*
- * Avispa μF - invoice generating software built on top of Avispa ECM
- * Copyright (C) 2023 Rafał Hiszpański
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Affero General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
- * GNU Affero General Public License for more details.
- *
- * You should have received a copy of the GNU Affero General Public License
- * along with this program.  If not, see <https://www.gnu.org/licenses/>.
- */
-
-package com.avispa.ecm.util.api.exception;
-
-import lombok.Getter;
-import org.springframework.http.HttpStatus;
-
-/**
- * @author Rafał Hiszpański
- */
-@Getter
-public class ApiException extends RuntimeException {
-    private final HttpStatus httpStatus;
-
-    public ApiException(HttpStatus httpStatus, String message) {
-        super(message);
-        this.httpStatus = httpStatus;
-    }
-
-    public ApiException(HttpStatus httpStatus, String message, Throwable cause) {
-        super(message, cause);
-        this.httpStatus = httpStatus;
-    }
-}
