+++ ecm-client-backend/src/main/resources/ecm-client-application.properties
@@ -1,5 +1,7 @@
 server.error.whitelabel.enabled=false
-server.error.path=/error
+server.error.include-stacktrace=always
+server.error.include-message=always
+#server.error.path=/error
 spring.thymeleaf.cache=false
 spring.cache.type=none
 
@@ -52,6 +54,8 @@ logging.level.org.springframework.jdbc.core.JdbcTemplate=DEBUG
 
 #---
 spring.config.activate.on-profile=prod
+
+server.error.include-stacktrace=never
 
 spring.jpa.hibernate.ddl-auto=validate
