+++ microf-custom/src/main/java/com/avispa/microf/model/bankaccount/BankAccountController.java
@@ -19,14 +19,15 @@
 package com.avispa.microf.model.bankaccount;
 
 import com.avispa.ecm.model.base.controller.BaseController;
-import com.avispa.microf.model.bankaccount.exception.BankAccountInUseException;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.UUID;
 
@@ -51,7 +52,7 @@ public class BankAccountController extends BaseController<BankAccount, BankAccou
         } catch(DataIntegrityViolationException e) {
             String errorMessage = String.format("Bank account '%s' is in use.", getService().findById(id).getObjectName());
             log.error(errorMessage, e);
-            throw new BankAccountInUseException(errorMessage);
+            throw new ResponseStatusException(HttpStatus.BAD_REQUEST, errorMessage);
         }
     }
 }
