+++ microf-custom/src/test/java/com/avispa/microf/model/invoice/service/InvoiceServiceIntegrationTest.java
@@ -23,10 +23,10 @@ import com.avispa.ecm.model.EcmObjectRepository;
 import com.avispa.ecm.model.configuration.context.ContextService;
 import com.avispa.ecm.model.configuration.template.Template;
 import com.avispa.ecm.model.content.Content;
-import com.avispa.ecm.model.error.ResourceNotFoundException;
 import com.avispa.ecm.model.format.Format;
 import com.avispa.ecm.model.format.FormatRepository;
 import com.avispa.ecm.service.rendition.RenditionService;
+import com.avispa.ecm.util.error.exception.ResourceNotFoundException;
 import com.avispa.ecm.util.exception.EcmException;
 import com.avispa.microf.model.bankaccount.BankAccount;
 import com.avispa.microf.model.customer.Customer;
